ESQL: add AbstractHashJoin base for hash joins
